Change in nomenclature
With this commit the terms `to` and `from` are changed to `recipient`
and `sender`, respectively. This was done due to Java method names
reading strangely if they are based on prepositions and not nouns. For
example, the method `addTo` is not as clear as `addRecipient` since it
is not obvious that we are here using the word "to" as a noun.

Unfortunately, this breaks "name compatibility" with the REST API. A
paragraph has been added to the tutorial highlighting this in case a
user wish to apply something from the REST API in their Java code.